Hito 4 aun no concluido

+++ app/Http/Controllers/ProjectController.php
@@ -9,20 +9,26 @@ Class ProjectController extends Controller
     // Mostrar lista de proyectos
     public function index()
     {
-        $projects = Project::latest()->get();
+        $projects = auth()->user()->projects()->latest()->get();
         return view('projects.index', compact('projects'));
     }
 
     // Guardar nuevo proyecto
-    public function store(Request $request)
+    public function store(Request $request, Project $project)
     {
+        
         $request->validate([
             'name' => 'required|string|max:255',
             'description' => 'nullable|string',
             'priority' => 'nullable|in:baja,media,alta'
         ]);
+        
 
-        Project::create(['name'=>$request->name, 'description'=>$request->description, 'priority'=>$request->priority]);
+        auth()->user()->projects()->create([
+            'name' => $request->name,
+            'description' => $request->description,
+            'priority' => $request->priority ?? 'media',
+        ]);
 
         return redirect('projects')->with('success', 'Proyecto creado exitosamente.');
     }
@@ -30,17 +36,20 @@ Class ProjectController extends Controller
     // Formulario para editar proyecto
     public function edit(Project $project)
     {
+            abort_if($project->user_id !== auth()->id(), 403);
             return view('projects.edit', compact('project'));
     }
 
     // Actualizar proyecto
     public function update(Request $request, Project $project)
     {
+        abort_if($project->user_id !== auth()->id(), 403);
         $request->validate([
             'name' => 'required|string|max:255',
             'description' => 'nullable|string',
         ]);
 
+        abort_if($project->user_id !== auth()->id(), 403);
         $project->update(['name'=>$request->name, 'description'=>$request->description]);
 
         return redirect('/projects')->with('success', 'Proyecto actualizado exitosamente.');
@@ -49,6 +58,7 @@ Class ProjectController extends Controller
     // Eliminar proyecto
     public function destroy(Project $project)
     {
+        abort_if($project->user_id !== auth()->id(), 403);
         $project->delete();
         return redirect('/projects')->with('success', 'Proyecto eliminado exitosamente.');
     }
